hello

+++ libft/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: rtamouss <rtamouss@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/01 17:44:31 by rtamouss          #+#    #+#             */
-/*   Updated: 2023/11/01 18:09:58 by rtamouss         ###   ########.fr       */
+/*   Updated: 2023/11/02 17:45:25 by rtamouss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(char *s)
+#include "libft.h"
+
+size_t	ft_strlen(char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
